model/requirements.txt
sentence-transformers
transformers
torch
pdfminer.six
python-docx
scikit-learn
numpy
